+++ libs/designer/src/lib/ui/panel/connectionsPanel/createConnection/_test__/createConnection.spec.tsx
@@ -1,9 +1,12 @@
+import { MockHttpClient } from '../../../../../__test__/mock-http-client';
 import { CreateConnection, type CreateConnectionProps } from '../createConnection';
 import { UniversalConnectionParameter } from '../formInputs/universalConnectionParameter';
 import {
   InitConnectionParameterEditorService,
   type IConnectionParameterEditorService,
   type IConnectionParameterInfo,
+  InitConnectionService,
+  StandardConnectionService,
 } from '@microsoft/designer-client-services-logic-apps';
 import type { ConnectionParameter, ConnectionParameterSets } from '@microsoft/utils-logic-apps';
 import React, { type ReactElement } from 'react';
@@ -11,6 +14,24 @@ import * as ReactShallowRenderer from 'react-test-renderer/shallow';
 
 describe('ui/createConnection', () => {
   let renderer: ReactShallowRenderer.ShallowRenderer;
+
+  const httpClient = new MockHttpClient();
+  InitConnectionService(
+    new StandardConnectionService({
+      apiVersion: '2018-07-01-preview',
+      baseUrl: '/baseUrl',
+      httpClient,
+      apiHubServiceDetails: {
+        apiVersion: '2018-07-01-preview',
+        baseUrl: '/baseUrl',
+        subscriptionId: '',
+        resourceGroup: '',
+        location: '',
+        httpClient,
+      },
+      readConnections: jest.fn(),
+    })
+  );
 
   beforeEach(() => {
     renderer = ReactShallowRenderer.createRenderer();
